Уличшения кода и Makefile
