+++ ping_pong/game/bot.cpp
@@ -1,3 +1,3 @@
-#include "gameobjects.hpp"
+#include "gameobj/player.hpp"
 
-int player::bot(float posY) { return posY - this->height / 2; }
+int player::bot(float posY) const { return posY - this->height / 2; }
