+++ ping_pong/game/screen.hpp
@@ -8,10 +8,9 @@ class screen {
     screen();
     ~screen();
     void draw(size_t frame_time = 0);
-    void add(int posY, int posX, std::vector<std::vector<char>>);
+    void add(int posY, int posX, const std::vector<std::vector<char>> &);
     void reset(char fill);
     scrsize getGameSize() const;
-    scrsize getScreenSize() const;
 
   private:
     std::vector<std::vector<char>> screen_vector;
